+++ src/views/explore/catalogs.tsx
@@ -2,15 +2,12 @@ import fetchApi from "~/shared/lib/api/fetch";
 import DetailsCard from "~/widgets/details-card";
 import GridContainer from "~/widgets/grid-container";
 
-import FavoriteCatalog from "./favorite-catalogs";
-
 export default async function Catalogs() {
   const catalogs = await fetchApi("/catalogs/valid");
   return (
     <div className="p-3">
       <h1 className="text-2xl font-semibold tracking-tight flex gap-2 items-start">
         <p>Catalogs</p>
-        <FavoriteCatalog />
       </h1>
       <div className="w-full pt-7">
         <GridContainer>