+++ src/widgets/app-sidebar.tsx
@@ -1,8 +1,10 @@
 "use client";
 
+import { useLiveQuery } from "dexie-react-hooks";
 import {
   Archive,
   BookOpen,
+  ChevronRight,
   Clock8,
   Compass,
   History,
@@ -10,9 +12,17 @@ import {
 } from "lucide-react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
+import { useState } from "react";
 
 import { useAuth } from "~/features/auth/context-provider";
+import { indexedDB } from "~/shared/lib/api/dexie";
 import { cn } from "~/shared/lib/tailwind-merge";
+import {
+  Collapsible,
+  CollapsibleContent,
+  CollapsibleTrigger,
+} from "~/shared/ui/collapsible";
+import { HeartListIcon } from "~/shared/ui/icons";
 import NextUpdate from "~/views/public-catalog/next-update";
 
 import { Avatar, AvatarFallback, AvatarImage } from "../shared/ui/avatar";
@@ -23,10 +33,15 @@ import {
   SidebarFooter,
   SidebarGroup,
   SidebarGroupContent,
+  SidebarGroupLabel,
   SidebarHeader,
   SidebarMenu,
+  SidebarMenuBadge,
   SidebarMenuButton,
   SidebarMenuItem,
+  SidebarMenuSub,
+  SidebarMenuSubItem,
+  SidebarSeparator,
   useSidebar,
 } from "../shared/ui/sidebar";
 import AuthButton from "./auth-buttons";
@@ -52,11 +67,88 @@ export default function AppSidebar() {
       <SidebarContent>
         <UserGroup />
         <ExploreGroup />
+        <SidebarSeparator />
+        <LocalGroup />
       </SidebarContent>
       <SidebarFooter>
         <FooterNextUpdate />
       </SidebarFooter>
     </Sidebar>
+  );
+}
+
+function LocalGroup() {
+  const { setOpenMobile } = useSidebar();
+  const [isOpen, setIsOpen] = useState(false);
+
+  const favoriteCatalogs =
+    useLiveQuery(() => indexedDB["favorites"].toArray(), []) ?? [];
+
+  return (
+    <Collapsible
+      open={isOpen}
+      onOpenChange={setIsOpen}
+      className="group/collapsible"
+    >
+      <SidebarGroup>
+        <SidebarGroupLabel
+          asChild
+          className={cn(
+            "group/label text-sm",
+            "w-full justify-start px-2",
+            "hover:bg-primary/5 hover:text-primary/80"
+          )}
+        >
+          <CollapsibleTrigger
+            className={cn(
+              "flex items-center gap-2 dark:text-white text-[#18181B]",
+              "group-data-[state=open]/collapsible:bg-primary/20 group-data-[state=open]/collapsible:dark:text-white"
+            )}
+          >
+            <HeartListIcon className="mr-2 h-4 w-4" />
+            Favorite Catalogs
+            <ChevronRight className="ml-auto transition-transform group-data-[state=open]/collapsible:rotate-90" />
+          </CollapsibleTrigger>
+        </SidebarGroupLabel>
+        <CollapsibleContent>
+          <SidebarGroupContent>
+            <SidebarMenu>
+              {favoriteCatalogs.map((item) => (
+                <SidebarMenuItem key={item.id}>
+                  <SidebarMenuSub>
+                    <SidebarMenuSubItem>
+                      <SidebarMenuButton
+                        onClick={() => setOpenMobile(false)}
+                        className={cn(
+                          "px-0",
+                          "data-[active=true]:bg-primary/20 data-[active=true]:dark:text-white",
+                          "data-[state=open]:hover:bg-transparent",
+                          "hover:bg-transparent"
+                        )}
+                        asChild
+                      >
+                        <Link href={`/c/${item.id}`}>
+                          <Button
+                            onClick={() => setIsOpen(false)}
+                            variant="ghost"
+                            className={cn(
+                              "w-full justify-start px-2",
+                              "hover:bg-primary/5 hover:text-primary/80"
+                            )}
+                          >
+                            {item.title}
+                          </Button>
+                        </Link>
+                      </SidebarMenuButton>
+                    </SidebarMenuSubItem>
+                  </SidebarMenuSub>
+                </SidebarMenuItem>
+              ))}
+            </SidebarMenu>
+          </SidebarGroupContent>
+        </CollapsibleContent>
+      </SidebarGroup>
+    </Collapsible>
   );
 }
 
@@ -130,6 +222,8 @@ function UserGroup() {
 
 function ExploreGroup() {
   const { setOpenMobile } = useSidebar();
+  const existingVideos =
+    useLiveQuery(() => indexedDB["watch-later"].toArray()) ?? [];
 
   const exploreItems = [
     { icon: Compass, label: "Explore", path: "/explore", shortPath: "/e/" },
@@ -146,6 +240,7 @@ function ExploreGroup() {
       shortPath: "/a/",
     },
     {
+      badge: existingVideos.length,
       icon: Clock8,
       label: "Watch later",
       path: "/explore/watch-later",
@@ -194,6 +289,9 @@ function ExploreGroup() {
                     </Button>
                   </Link>
                 </SidebarMenuButton>
+                {item.badge ? (
+                  <SidebarMenuBadge>{item.badge}</SidebarMenuBadge>
+                ) : null}
               </SidebarMenuItem>
             );
           })}