+++ src/widgets/youtube/player.tsx
@@ -12,42 +12,9 @@ export default function YoutubePlayer(
   props: VideoData & { enableJsApi: boolean }
 ) {
   const { enableJsApi, ...video } = props;
-  const playerRef = useRef<YT.Player | null>(null);
   const containerRef = useRef<HTMLDivElement | null>(null);
   const trackingRef = useRef<NodeJS.Timeout | null>(null);
-
-  async function loadIFrameElement() {
-    if (!enableJsApi) {
-      return;
-    }
-
-    if (!playerRef.current) {
-      playerRef.current = await (
-        containerRef.current?.querySelector("lite-youtube") as any
-      )?.getYTPlayer();
-
-      const playing = await indexedDB["history"].get(video.videoId);
-
-      if (playing?.videoId === video.videoId) {
-        playerRef.current?.seekTo(playing.duration, true);
-      }
-
-      // Set up player event listeners
-      playerRef.current?.addEventListener("onStateChange", (event) => {
-        const playerState = window.YT.PlayerState;
-
-        switch (event.data) {
-          case playerState.PLAYING:
-            startTracking();
-            break;
-          case playerState.PAUSED:
-          case playerState.ENDED:
-            stopTracking();
-            break;
-        }
-      });
-    }
-  }
+  const playerRef = useRef<YT.Player | null>(null);
 
   const startTracking = () => {
     if (trackingRef.current) return;
@@ -69,7 +36,7 @@ export default function YoutubePlayer(
       };
 
       await indexedDB["history"].put(payload);
-    }, 3_000);
+    }, 2_000);
   };
 
   const stopTracking = () => {
@@ -79,12 +46,47 @@ export default function YoutubePlayer(
     }
   };
 
+  function _onStateChange(event: YT.OnStateChangeEvent) {
+    const playerState = window.YT.PlayerState;
+
+    switch (event.data) {
+      case playerState.PLAYING:
+        startTracking();
+        break;
+      case playerState.PAUSED:
+      case playerState.ENDED:
+        stopTracking();
+        break;
+    }
+  }
+
+  async function loadIFrameElement() {
+    if (!enableJsApi) {
+      return;
+    }
+
+    if (!playerRef.current) {
+      playerRef.current = await (
+        containerRef.current?.querySelector("lite-youtube") as any
+      )?.getYTPlayer();
+
+      const playing = await indexedDB["history"].get(video.videoId);
+
+      if (playing?.videoId === video.videoId) {
+        playerRef.current?.seekTo(playing.duration, true);
+      }
+
+      // Set up player event listeners
+      playerRef.current?.addEventListener("onStateChange", _onStateChange);
+    }
+  }
+
   useEffect(() => {
     return () => stopTracking();
   }, []);
 
   return (
-    <div ref={containerRef} onMouseDown={() => loadIFrameElement()}>
+    <div ref={containerRef} onMouseDown={loadIFrameElement}>
       <YouTubeEmbed
         params={enableJsApi ? IframeParams + "&enablejsapi=1" : IframeParams}
         videoid={video.videoId}